pyproject.toml
[tool.poetry]
name = "x-thread-extractor"
version = "0.1.0"
description = "A tool to extract main threads from X (Twitter) conversations"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.31.0"
python-dotenv = "^1.0.0"
click = "^8.1.7"
pydantic = "^2.5.2"
openai = "^1.3.0"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.12.0"
mypy = "^1.7.0"
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
x-extract = "x_thread_extractor.cli:main"
